+++ src/libdieharder/countx.c
@@ -15,6 +15,7 @@
 #include <float.h>
 #include <time.h>
 #include <R_ext/Print.h>
+#include <stdint.h>
 
 typedef  unsigned char      u1;
 typedef  unsigned long      u4;
@@ -233,7 +234,7 @@ int main_countx( int argc, char **argv)
   free(data);
 
   time(&z);
-  Rprintf("number of seconds: %6lu\n", (size_t)(z-a));
+  Rprintf("number of seconds: %6u\n", (uint32_t)(z-a));
 
   return 0;
